requirements.txt
Django==4.2.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
python-dotenv==1.0.0
requests==2.28.2
djangorestframework==3.14.0
mysql-connector==2.2.9